Dovrsen clean code za premestanje opreme i izmenjeni dijagrami sekvence i aktivnosti za funkcionalnosti upravnika

+++ Projekat/bolnica/bolnica/Forms/Upravnik/FormSastojci.xaml.cs
@@ -69,16 +69,47 @@ namespace Bolnica.Forms.Upravnik
         {
             if (dataGridSastojci.SelectedItems.Count > 0)
             {
-                var selectedItems = dataGridSastojci.SelectedItems;
-                List<Sastojak> sastojciZaBrisanje = new List<Sastojak>();
-                foreach (Sastojak s in selectedItems)
+                string sMessageBoxText = "";
+                if (dataGridSastojci.SelectedItems.Count > 1)
+                    sMessageBoxText = "Da li ste sigurni da želite da obrišete izabrane sastojke?";
+                else
+                    sMessageBoxText = "Da li ste sigurni da želite da obrišete sastojak " + ((Sastojak)dataGridSastojci.SelectedItem).Naziv + "?";
+                string sCaption = "Brisanje opreme";
+
+                MessageBoxButton btnMessageBox = MessageBoxButton.YesNo;
+                MessageBoxImage icnMessageBox = MessageBoxImage.Warning;
+
+                MessageBoxResult rsltMessageBox = MessageBox.Show(sMessageBoxText, sCaption, btnMessageBox, icnMessageBox);
+
+
+                if (rsltMessageBox == MessageBoxResult.Yes)
                 {
-                    sastojciZaBrisanje.Add(s);
-                    storage.Delete(s);
+                    var selectedItems = dataGridSastojci.SelectedItems;
+                    List<Sastojak> sastojciZaBrisanje = new List<Sastojak>();
+                    foreach (Sastojak s in selectedItems)
+                    {
+                        sastojciZaBrisanje.Add(s);
+                        storage.Delete(s);
+                    }
+                    foreach (Sastojak s in sastojciZaBrisanje)
+                    {
+                        Sastojci.Remove(s);
+                    }
                 }
-                foreach (Sastojak s in sastojciZaBrisanje)
+            }
+        }
+
+        private void OnKeyDownHandler(object sender, KeyEventArgs e)
+        {
+            if (e.Key == Key.Return)
+            {
+                Sastojci.Clear();
+                foreach(Sastojak s in storage.GetAll())
                 {
-                    Sastojci.Remove(s);
+                    if(s.Naziv.ToLower().StartsWith(txtSearch.Text.ToLower()))
+                    {
+                        Sastojci.Add(s);
+                    }
                 }
             }
         }
